First shooting version!
In theory, the shooting system is functional. I will work in the following days at changing the implementation of key input to suite the rest of the project, but for now it is all we need. Also I will look into progressive spray, it will be included in the version for the second deadline.